requirements.txt
# AMR Multi-Agent Framework — Dependencies
# Install with: pip install -r requirements.txt

# ── Dashboard ──────────────────────────────────────────────────────────
streamlit>=1.32.0          # Web dashboard (Fleet Map, Analytics, Scenarios)
plotly>=5.18.0             # Interactive charts (battery, Q-table heatmap, radar)

# ── Data & ML ──────────────────────────────────────────────────────────
numpy>=1.24.0              # Q-table (Energy Agent), feature vectors (Sentinel)
scikit-learn>=1.3.0        # IsolationForest anomaly detection (Sentinel Agent)
pandas>=2.0.0              # DataFrames for dashboard tables and CSV export

# ── LLM (optional — free tier) ─────────────────────────────────────────
groq>=0.4.0                # Groq API client -> LLaMA3-70B for path negotiation
                           # and incident reports. Get free key: console.groq.com
                           # Framework runs without this — agents fall back to
                           # rule-based logic if GROQ_API_KEY is not set.

# ── Config & Utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.0       # Load .env file (GROQ_API_KEY, NUM_AMRS, etc.)

# ── Testing (optional) ─────────────────────────────────────────────────
pytest>=7.0.0              # Only needed if using pytest directly.
                           # The built-in runner works without it:
                           # python run_tests.py

# ── Standard Library (no install needed) ───────────────────────────────
# asyncio      — async event loop       (built-in Python 3.10+)
# sqlite3      — persistence layer      (built-in)
# dataclasses  — data models            (built-in Python 3.7+)
# collections  — deque rolling buffers  (built-in)
# logging      — structured logging     (built-in)
# pathlib      — file paths             (built-in)
# csv / json   — export formats         (built-in)

# ── Python Version ─────────────────────────────────────────────────────
# Requires Python >= 3.10
# Tested on Python 3.10, 3.11, 3.12